Fix VTKWriter
In the merge with Alexis work, one place was missed for the
float->double position upgrade, resulting in broken VTKs being produced.

Trivial fix.

+++ src/VTKWriter.cc
@@ -306,7 +306,7 @@ void VTKWriter::write(uint numParts, const double4 *pos, const float4 *vel,
 		}
 	}
 
-	numbytes=sizeof(float)*3*numParts;
+	numbytes=sizeof(double)*3*numParts;
 
 	// position
 	fwrite(&numbytes, sizeof(numbytes), 1, fid);
